Migrated 'Observer' class to 'murfey.util.client' as well
