+++ src/murfey/util/__init__.py
@@ -1,12 +1,10 @@
 from __future__ import annotations
 
-import asyncio
-import inspect
 import logging
 from pathlib import Path
 from queue import Queue
 from threading import Thread
-from typing import Awaitable, Callable, Optional
+from typing import Optional
 from uuid import uuid4
 
 from werkzeug.utils import secure_filename
@@ -63,75 +61,6 @@ def posix_path(path: Path) -> str:
         posix_path = "/".join(path_parts)
         return posix_path
     return str(path)
-
-
-class Observer:
-    """
-    A helper class implementing the observer pattern supporting both
-    synchronous and asynchronous notification calls and both synchronous and
-    asynchronous callback functions.
-    """
-
-    # The class here should be derived from typing.Generic[P]
-    # with P = ParamSpec("P"), and the notify/anotify functions should use
-    # *args: P.args, **kwargs: P.kwargs.
-    # However, ParamSpec is Python 3.10+ (PEP 612), so we can't use that yet.
-
-    def __init__(self):
-        self._listeners: list[Callable[..., Awaitable[None] | None]] = []
-        self._secondary_listeners: list[Callable[..., Awaitable[None] | None]] = []
-        self._final_listeners: list[Callable[..., Awaitable[None] | None]] = []
-        super().__init__()
-
-    def subscribe(
-        self,
-        fn: Callable[..., Awaitable[None] | None],
-        secondary: bool = False,
-        final: bool = False,
-    ):
-        if final:
-            self._final_listeners.append(fn)
-        elif secondary:
-            self._secondary_listeners.append(fn)
-        else:
-            self._listeners.append(fn)
-
-    async def anotify(
-        self, *args, secondary: bool = False, final: bool = False, **kwargs
-    ) -> None:
-        awaitables: list[Awaitable] = []
-        listeners = (
-            self._secondary_listeners
-            if secondary
-            else self._final_listeners if final else self._listeners
-        )
-        for notify_function in listeners:
-            result = notify_function(*args, **kwargs)
-            if result is not None and inspect.isawaitable(result):
-                awaitables.append(result)
-        if awaitables:
-            await self._await_all(awaitables)
-
-    @staticmethod
-    async def _await_all(awaitables: list[Awaitable]):
-        for awaitable in asyncio.as_completed(awaitables):
-            await awaitable
-
-    def notify(
-        self, *args, secondary: bool = False, final: bool = False, **kwargs
-    ) -> None:
-        awaitables: list[Awaitable] = []
-        listeners = (
-            self._secondary_listeners
-            if secondary
-            else self._final_listeners if final else self._listeners
-        )
-        for notify_function in listeners:
-            result = notify_function(*args, **kwargs)
-            if result is not None and inspect.isawaitable(result):
-                awaitables.append(result)
-        if awaitables:
-            asyncio.run(self._await_all(awaitables))
 
 
 class Processor:
