+++ src/murfey/util/client.py
@@ -4,15 +4,19 @@ configuration, communicate with the backend server using the correct credentials
 and set default directories to work with.
 """
 
+from __future__ import annotations
+
+import asyncio
 import configparser
 import copy
+import inspect
 import json
 import logging
 import os
 import shutil
 from functools import lru_cache, partial
 from pathlib import Path
-from typing import Callable, Optional, Union
+from typing import Awaitable, Callable, Optional, Union
 from urllib.parse import ParseResult, urlparse, urlunparse
 
 import requests
@@ -169,3 +173,72 @@ def set_default_acquisition_output(
         if _check_dict_structure(settings, settings_copy):
             with open(p, "w") as sf:
                 json.dump(settings_copy, sf)
+
+
+class Observer:
+    """
+    A helper class implementing the observer pattern supporting both
+    synchronous and asynchronous notification calls and both synchronous and
+    asynchronous callback functions.
+    """
+
+    # The class here should be derived from typing.Generic[P]
+    # with P = ParamSpec("P"), and the notify/anotify functions should use
+    # *args: P.args, **kwargs: P.kwargs.
+    # However, ParamSpec is Python 3.10+ (PEP 612), so we can't use that yet.
+
+    def __init__(self):
+        self._listeners: list[Callable[..., Awaitable[None] | None]] = []
+        self._secondary_listeners: list[Callable[..., Awaitable[None] | None]] = []
+        self._final_listeners: list[Callable[..., Awaitable[None] | None]] = []
+        super().__init__()
+
+    def subscribe(
+        self,
+        fn: Callable[..., Awaitable[None] | None],
+        secondary: bool = False,
+        final: bool = False,
+    ):
+        if final:
+            self._final_listeners.append(fn)
+        elif secondary:
+            self._secondary_listeners.append(fn)
+        else:
+            self._listeners.append(fn)
+
+    async def anotify(
+        self, *args, secondary: bool = False, final: bool = False, **kwargs
+    ) -> None:
+        awaitables: list[Awaitable] = []
+        listeners = (
+            self._secondary_listeners
+            if secondary
+            else self._final_listeners if final else self._listeners
+        )
+        for notify_function in listeners:
+            result = notify_function(*args, **kwargs)
+            if result is not None and inspect.isawaitable(result):
+                awaitables.append(result)
+        if awaitables:
+            await self._await_all(awaitables)
+
+    @staticmethod
+    async def _await_all(awaitables: list[Awaitable]):
+        for awaitable in asyncio.as_completed(awaitables):
+            await awaitable
+
+    def notify(
+        self, *args, secondary: bool = False, final: bool = False, **kwargs
+    ) -> None:
+        awaitables: list[Awaitable] = []
+        listeners = (
+            self._secondary_listeners
+            if secondary
+            else self._final_listeners if final else self._listeners
+        )
+        for notify_function in listeners:
+            result = notify_function(*args, **kwargs)
+            if result is not None and inspect.isawaitable(result):
+                awaitables.append(result)
+        if awaitables:
+            asyncio.run(self._await_all(awaitables))
